Challenge #2
Write a Program, to input a number from the user, and check whether the input number is a Perfect Number, or not.

The condition is that the input number should be greater than 2 and less than 5000, and you have to keep taking the input again and again, until the user enters the correct input.