Challenge #5
Write a program, to find out whether an input number is palindromic or not. A palindromic number, if reversed, yields the same number. For example, 121. If you reverse 121, you also get 121. Thus, 121 is palindromic. 1234 reversed is 4321, and hence not palindromic.

Input constraint: The input number should be greater than 0. If the user doesn't enter a number greater than 0, they should be asked for the number again (use a while / do-while loop).